bi-engine/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.0.1
